Refactor: move Chest to `src` module, add `image` param; add type hints
